+++ src/Data/Decklists.ts
@@ -1,18 +1,19 @@
 import { Decklist, GET_TOURNAMENT_ID, MOCK_DECKLIST_FROM_RAW, RawTournamentResults, TournamentResults } from "./Interfaces";
-import { AEGIS_20250817 } from "./decklists/Aegis";
-import { BJO_20250831, CQO_20250831, GZO_20250824, SFC_20250803 } from "./decklists/China";
-import { EXPERT_DECKLISTS } from "./decklists/ExpertDecklists";
-import { RFC_20250809 } from "./decklists/France";
-import { MCW_20250807 } from "./decklists/Italy";
-import { LUX_20250816 } from "./decklists/LuxuryRiftboundChampionSeries";
-import { RMW_20250809, RMW_20250802, RMW_20250726 } from "./decklists/RiftboundMetaWeekly";
-import { RLL_20250905, RLL_20250822, RLL_20250823, RLT_20250808 } from "./decklists/Riftlab";
-import { TNF_20250904, TNF_20250828, TNF_20250807, TNF_20250731, TNF_20250814, TNF_20250821 } from "./decklists/ThursdayNightFights";
+import { AEGIS_20250817 } from "./Decklists/Aegis";
+import { HZO_20250914, BJO_20250831, CQO_20250907, GZO_20250824, SFC_20250803 } from "./Decklists/China";
+import { EXPERT_DECKLISTS } from "./Decklists/PinnedDecklists";
+import { RFC_20250809 } from "./Decklists/France";
+import { MCW_20250807 } from "./Decklists/Italy";
+import { LUX_20250816 } from "./Decklists/LuxuryRiftboundChampionSeries";
+import { RMW_20250809, RMW_20250802, RMW_20250726 } from "./Decklists/RiftboundMetaWeekly";
+import { RLL_20250919, RLL_20250905, RLL_20250822, RLL_20250823, RLT_20250808 } from "./Decklists/Riftlab";
+import { TNF_20250918, TNF_20250911, TNF_20250904, TNF_20250828, TNF_20250807, TNF_20250731, TNF_20250814, TNF_20250821 } from "./Decklists/ThursdayNightFights";
 
 const RAW_TOURNAMENT_RESULTS: RawTournamentResults[] = [
-    CQO_20250831, RLL_20250905, TNF_20250904, BJO_20250831, TNF_20250828, GZO_20250824, RLL_20250823, RLL_20250822, TNF_20250821, AEGIS_20250817, LUX_20250816, TNF_20250814, RMW_20250809, RFC_20250809, RLT_20250808, MCW_20250807, TNF_20250807, SFC_20250803, RMW_20250802, TNF_20250731, RMW_20250726
+    RLL_20250919, TNF_20250918, HZO_20250914, TNF_20250911, CQO_20250907, RLL_20250905, TNF_20250904, BJO_20250831, TNF_20250828, GZO_20250824, RLL_20250823, RLL_20250822, TNF_20250821, AEGIS_20250817, LUX_20250816, TNF_20250814, RMW_20250809, RFC_20250809, RLT_20250808, MCW_20250807, TNF_20250807, SFC_20250803, RMW_20250802, TNF_20250731, RMW_20250726
 ];
 
+let unknownPlayerCount = 0;
 export let TOURNAMENT_RESULTS: TournamentResults[] = [];
 for(let raw of RAW_TOURNAMENT_RESULTS) {
     let tournamentResults: TournamentResults = {...raw} as TournamentResults;
@@ -22,6 +23,9 @@ for(let raw of RAW_TOURNAMENT_RESULTS) {
             decklist.date = tournamentResults.date;
             decklist.placing = placing.placing;
             decklist.tournId = GET_TOURNAMENT_ID(tournamentResults.abbrName, tournamentResults.date);
+            if(decklist.username === "") {
+                decklist.username = `*Unknown${unknownPlayerCount++}*`;
+            }
         }
     }
     TOURNAMENT_RESULTS.push(tournamentResults);
@@ -120,10 +124,11 @@ const archetypeLegendDict: {[archetype: string]: string} = {
     "Lux Control": "OGS-021",
     "Garen Midrange": "OGS-023",
     "Kai'Sa": "OGN-247",
+    "Kai'Sa Control": "OGN-247",
     "Volibear Ramp": "OGN-249",
     "Jinx Aggro": "OGN-251",
     "Darius": "OGN-253",
-    "Ahri Tempo": "OGN-255",
+    "Ahri": "OGN-255",
     "Lee Sin Midrange": "OGN-257",
     "Yasuo Midrange": "OGN-259",
     "Leona Midrange": "OGN-261",
@@ -142,8 +147,8 @@ export const ALL_ARCHETYPES: string[] = Object.keys(archetypeLegendDict).sort();
 
 export const ARCHETYPE_TIERS = [
     ["Kai'Sa", "Master Yi Midrange"],
-    ["Miss Fortune Ramp", "Sett Midrange", "Darius"],
-    ["Viktor", "Annie Tempo", "Teemo", "Ahri Tempo", "Master Yi Ramp"],
+    ["Miss Fortune Ramp", "Darius", "Sett Midrange", "Master Yi Ramp"],
+    ["Viktor", "Annie Tempo", "Teemo", "Ahri", "Kai'Sa Control"],
     ["Lee Sin Midrange", "Sett Ramp", "Volibear Ramp", "Lux Control", "Jinx Aggro", "Leona Midrange", "Yasuo Midrange"],
     ["Garen Midrange"],
 ];
@@ -162,81 +167,3 @@ export function GET_CASED_ARCHETYPE(archetype: string): string {
     }
     return archetypeCasedNames[archetypeLower];
 }
-
-// {
-//     "Viktor Swarm Control": [
-//         {
-//             username: "PochiPoom",
-//             link: "https://piltoverarchive.com/decks/view/3889db57-5e18-4b15-9303-a2f6b3ab9ef6",
-//             cardAmounts: [
-//                 // LEGEND
-//                 {id: "OGN-265-1", count: 1},
-//                 // CHOSEN CHAMPION
-//                 {id: "OGN-246-1", count: 1},
-//                 // MAIN DECK
-//                 {id: "OGN-095-1", count: 3},
-//                 {id: "OGN-104-1", count: 2},
-//                 {id: "OGN-108-1", count: 2},
-//                 {id: "OGN-213-1", count: 3},
-//                 {id: "OGN-093-1", count: 3},
-//                 {id: "OGN-210-1", count: 3},
-//                 {id: "OGN-209-1", count: 2},
-//                 {id: "OGN-241-1", count: 2},
-//                 {id: "OGN-211-1", count: 3},
-//                 {id: "OGN-094-1", count: 3},
-//                 {id: "OGN-218-1", count: 3},
-//                 {id: "OGN-208-1", count: 3},
-//                 {id: "OGN-233-1", count: 2},
-//                 {id: "OGN-105-1", count: 2},
-//                 {id: "OGN-116-1", count: 2},
-//                 {id: "OGN-122-1", count: 1},
-//                 // BATTLEFIELDS
-//                 {id: "OGN-294-1", count: 1},
-//                 {id: "OGN-284-1", count: 1},
-//                 {id: "OGN-295-1", count: 1},
-//                 // RUNE DECK
-//                 {id: "OGN-214-1", count: 6},
-//                 {id: "OGN-089-1", count: 6},
-//                 // SIDEBOARD
-//                 {id: "OGN-224-1", count: 2},
-//                 {id: "OGN-238-1", count: 3},
-//                 {id: "OGN-232-1", count: 3},
-//             ]
-//         },
-//         {
-//             username: "Zach (Riftbound Meta)",
-//             link: "https://piltoverarchive.com/decks/view/7209fd9d-beb9-4afb-8f9c-da7124a0c2a3",
-//             cardAmounts: [
-//                 // LEGEND
-//                 {id: "OGN-265-1", count: 1},
-//                 // CHOSEN CHAMPION
-//                 {id: "OGN-246-1", count: 1},
-//                 // MAIN DECK
-//                 {id: "OGN-095-1", count: 3},
-//                 {id: "OGN-103-1", count: 3},
-//                 {id: "OGN-213-1", count: 3},
-//                 {id: "OGN-210-1", count: 3},
-//                 {id: "OGN-266-1", count: 3},
-//                 {id: "OGN-209-1", count: 3},
-//                 {id: "OGN-096-1", count: 2},
-//                 {id: "OGN-222-1", count: 3},
-//                 {id: "OGN-211-1", count: 2},
-//                 {id: "OGN-218-1", count: 3},
-//                 {id: "OGN-208-1", count: 3},
-//                 {id: "OGN-083-1", count: 2},
-//                 {id: "OGN-233-1", count: 3},
-//                 {id: "OGN-234-1", count: 3},
-//                 // BATTLEFIELDS
-//                 {id: "OGN-294-1", count: 1},
-//                 {id: "OGN-284-1", count: 1},
-//                 {id: "OGN-285-1", count: 1},
-//                 // RUNE DECK
-//                 {id: "OGN-214-1", count: 8},
-//                 {id: "OGN-089-1", count: 4},
-//                 // SIDEBOARD
-//                 {id: "OGN-119-1", count: 3},
-//                 {id: "OGN-085-1", count: 2},
-//                 {id: "OGN-243-1", count: 3},
-//             ]
-//         },
-//     ],