+++ docs/RELEASE-NOTES.md
@@ -1,3 +1,8 @@
-[1.1.1] Use print id as id + fixes
-- Multi-colored cards
-- Follow import order conventions
+[1.2.0] Archetype Decklist Table Visual Update
+- update VDecklistTable to sticky + aesthetically
+- update TournamentResults to side scroll
+- VDecklistCard all same width
+- many more tournament results
+- support blank usernames
+- fix get ccategory bug from using wrong id
+- fix tts export using wrong id